requirements.txt
imageio==2.35.1
numpy==2.1.1
pillow==10.4.0
pyparsing==3.1.4
scipy==1.14.1
six==1.16.0
